Redesign create listing page
refs #47
